setting-up-receiving-user-data

+++ Microservices/AccountsMicroservice/AccountsBLL/Services/Implementations/FinancialAccountService.cs
@@ -1,6 +1,7 @@
 ﻿using Abp.Domain.Entities;
 using Accounts.BusinessLogic.Consumers;
 using Accounts.BusinessLogic.Models;
+using Accounts.BusinessLogic.Producers;
 using Accounts.BusinessLogic.Services.Interfaces;
 using Accounts.DataAccess.Entities;
 using Accounts.DataAccess.Settings;
@@ -14,29 +15,46 @@ namespace Accounts.BusinessLogic.Services.Implementations
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
         private readonly IMessageConsumer _consumer;
+        private readonly IMessageProducer _producer;
 
-        public FinancialAccountService(IUnitOfWork unitOfWork, IMapper mapper, IMessageConsumer consumer)
+        public FinancialAccountService(IUnitOfWork unitOfWork, IMapper mapper, 
+            IMessageConsumer consumer, IMessageProducer producer)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _consumer = consumer;
+            _producer = producer;
         }
 
         public async Task AddAsync(FinancialAccountModel addModel, CancellationToken cancellationToken)
         {
+            int userId = _consumer.ConsumeMessage(addModel.UserId);
+
+            if (userId == 0)
+            {
+                throw new Exception("The user is not logged in");
+            }
+
             var account = _mapper.Map<FinancialAccount>(addModel);
             await _unitOfWork.FinancialAccounts.AddAsync(account, cancellationToken);
 
             await _unitOfWork.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(int id, CancellationToken cancellationToken)
+        public async Task DeleteAsync(int userId, int id, CancellationToken cancellationToken)
         {
             var account = await _unitOfWork.FinancialAccounts.GetByIdAsync(id, cancellationToken);
 
             if (account == null)
             {
                 throw new EntityNotFoundException("Account not found");
+            }
+
+            int receivedUserId = _consumer.ConsumeMessage(account.UserId);
+
+            if (receivedUserId == 0 || userId != receivedUserId)
+            {
+                throw new Exception("The user is not logged in");
             }
 
             await _unitOfWork.FinancialAccounts.DeleteAsync(id, cancellationToken);
@@ -56,11 +74,18 @@ namespace Accounts.BusinessLogic.Services.Implementations
         public async Task<List<FinancialAccountModel>> GetAccountsByUserIdAsync(int userId,
             PaginationSettings paginationSettings, CancellationToken cancellationToken)
         {
-            int id = _consumer.ConsumeMessage(userId.ToString());
+            int id = _consumer.ConsumeMessage(userId);
+
+            if (id == 0)
+            {
+                throw new Exception("The user is not logged in");
+            }
 
             var accounts = await _unitOfWork.FinancialAccounts
                 .GetAccountsByUserIdAsync(id, paginationSettings, cancellationToken);
             var accountsList = _mapper.Map<List<FinancialAccountModel>>(accounts);
+
+            _producer.SendMessages(accountsList);   
 
             return accountsList;
         }
@@ -86,6 +111,13 @@ namespace Accounts.BusinessLogic.Services.Implementations
             if (account == null)
             {
                 throw new EntityNotFoundException("Account not found");
+            }
+
+            int userId = _consumer.ConsumeMessage(updateModel.UserId);
+
+            if (userId == 0)
+            {
+                throw new Exception("The user is not logged in");
             }
 
             var updateAccount = _mapper.Map<FinancialAccount>(updateModel);
